refactored main method to class
